Refactor Utilities into Utilities folder.
