+++ TrackerAppProject/TrackerApp/Tracker.cs
@@ -1,4 +1,5 @@
 ﻿using Spectre.Console;
+using TrackerApp.Utilities;
 
 namespace TrackerApp;
